Take the data file path from ${AQUARIA_DATA_PATH}, if present.
This patch allows the environment variable ${AQUARIA_DATA_PATH} to be used
to override the default path for data files (the executable's directory).
By running Aquaria as e.g.
    AQUARIA_DATA_PATH=/usr/games/lib/aquaria /usr/games/bin/aquaria
Aquaria will look in the specified directory for all its data files, so
the executable can be installed in a different place.

This currently only works on Linux (i.e., when BBGE_BUILD_UNIX is set),
but it could work for other systems as well with appropriate code in
Aquaria/Main.cpp (to get the environment variable or other system setting)
and Core::initPlatform() (to set the directory appropriately).