Improve keyboard/joystick control on the title screen.
This patch updates the title menu processing so that when a submenu
(such as the options menu or quit confirmation dialog) is closed under
keyboard or joystick control, the mouse cursor returns to the location
of the associated title menu node.  The cursor is _not_ moved when under
real mouse control, since the sudden jump in position could confuse the
user.  The patch also enables the "Replay Intro" node (the blue crystal)
to be selected under keyboard or joystick control.

This patch also tweaks the mod selector so that either of the action
buttons (the left- or right-mouse-button equivalent) can be used to
start a mod.

As part of this change, the engine exports a new script function
getInputMode() and associated INPUT_* constants, allowing scripts to
take actions based on the current input mode (mouse, keyboard, or
joystick).