Disallow changing to spirit form at warp points.
If Naija warps to another area while in spirit form, the form and world
type will become desynchronized in the new area -- Naija will show as a
spirit beacon and time will be frozen, but she can move around and sing
as usual, even changing to other forms.  Furthermore, if Naija uses the
spirit form song, the world will revert to normal but she will be unable
to move more than a certain distance (as is usual in spirit form).

Normally, Naija is blocked from entering warp nodes while in spirit
form.  However, if she rides a seahorse rightward into a warp node and
the last three notes she sings match the spirit form song, the song
will take effect immediately before the warp, leading to the problems
described above.  This patch works around the problem by disabling the
spirit form song when Naija is inside a warp node; the song symbol will
appear for an instant as the screen fades out, but the song itself will
not take effect.

This was reported on the Bit Blot forum, at (among other places):
http://www.bit-blot.com/forum/index.php?topic=890.0