Prevent the Mithalas boss from using an improper attack.
This patch fixes a variable name typo and related minor issues which
caused the Mithalas Cathedral boss to execute a "stomp" attack even when
the stomping foot would land on the pit area.  With this change, the
boss instead executes a "grab" attack under that condition (which will
generally hit Naija when she is in the area that triggers this logic).