Don't play the monkey damage sound effect for drowned monkeys.
This patch tweaks the script for the monkeys found in the Veil so that
the "screech" sound effect used for damage is not played when hitting a
monkey which has already drowned.

As a side effect, this patch also allows monkeys to take damage (and be
killed) during the "drowning" animation.  Damage sound effects were
already disabled during drowning, and remain so.