Keep ingredients in cooking slots if there are leftovers.
This patch implements a suggestion by Dustin Voss on the Aquaria mailing
list to speed up cooking of multiple recipes:  If the player has enough
stocked ingredients to cook more of the same recipe, the ingredients are
kept in the cooking slots, and the inventory slots are "bumped" (the
short expand/shrink animation) to show that one of each ingredient has
been used up.  This allows the player to simply select "Cook" several
times in succession to cook two or more of the same dish, provided she
has enough ingredients.