Fix OpenAL channel allocation handling.
This patch corrects the OpenAL setup logic to explicitly request the
number of channels expected by Aquaria, and to limit the number of
channels actually exported to the number returned by the OpenAL
implementation.  This fixes incorrect behavior (such as delayed sounds
or crashing) when the system defaults to less than 128 channels.