Implement a single-environment script execution model.
This patch implements a new script execution model in which all scripts
execute within a single Lua environment, rather than establishing a
separate environment for each script.  This model significantly reduces
memory usage, such that typical areas should not use more than around
1MB of memory (as compared to 10-20MB with the original model).

IMPORTANT NOTE:  The changes introduced here BREAK COMPATIBILITY with
the script interface provided by Aquaria v1.1.3 and earlier.  The
scripts distributed with such versions will not work with this code,
and scripts designed for this interface will not work with the game
executable from prior versions.

The new execution model requires some significant changes to existing
scripts in order for the scripts to function properly.  These changes
are documented in detail at the top of ScriptInterface.cpp; in summary:

- All scripts should include a "v = getVars()" line at the top of the
  script to obtain a reference to the instance-scope variable table.

- Scripts must not declare or use any global variables or functions,
  except for the "v" table and the specific interface functions called
  by the engine (init(), update(), and so on).  Constants and private
  helper functions can be declared with "local"; variables must be
  stored in the "v" table.

- Scripts must not call any functions or initialize any table variables
  outside a function; such code must go in init().

- Scripts must not call any interface functions (like init() or
  update()).  Common logic used by multiple interface functions should
  go into a helper function.

In order to help detect script errors, this patch also adds the
compile-time options "complainOnGlobalVar" and "complainOnUndefLocal",
which enable warnings via errorLog() when a script tries to get or set
a global variable or get an undefined local variable, respectively.
The warnings are disabled by default, but can be enabled by setting the
appropriate constant to "true" in Aquaria/ScriptInterface.cpp.