Limit the rate of splash effects.
This patch adds a rate limit to the splash effects (sound and visual)
used when Naija jumps out of or into the water.  In particular, this
avoids sound distortion which can occur when Naija swims along the
surface of the water, continually triggering splash effects.

This was reported on the Bit Blot forum, at:
http://www.bit-blot.com/forum/index.php?topic=1563.0