Remember the current menu page across area changes.
This patch ensures that the current menu page in the in-game menu,
along with the current page number in the food and treasure lists, is
maintained across area changes.  Previously, the menu would reset to
the songs page and page 1 of food and treasures each time a new area
was loaded.

Based on a patch supplied by <andrews05@gmail.com>.