Ensure that entities always recognize their songs properly.
This patch fixes a bug in the DeepWhale and LiCage entities in which the
song to activate the entity would not be recognized if immediately
preceded by singing part (but not all) of that song.  For example, given
the song A-B-C-D, the sequence of notes A-B-A-B-C-D would trigger this
bug: the second A is tested against the expected note C and fails, but
is not then checked against the first note of the song.