Prevent Li and the pet blaster from hitting seahorses and other non-targets.
This patch alters the collision logic for the energy shots fired by Li
and the pet blaster so that they do not collide with entities which are
explicitly set to ignore those attacks.  For example, this allows Li to
shoot through seahorses and hit enemies on the far side (just as Naija's
energy attack does), rather than hitting and potentially killing the
seahorse.