Shrink minimap icons when the corresponding locations are distant.
This patch tweaks the display of the warp point (white/gold circle),
save point (red circle), and cooking point icons on the minimap display
so that they shrink in size as the distance to the corresponding point
increases.  Based on a suggestion from "breversa" on the Aquaria mailing
list.

This patch also includes several optimizations to the minimap rendering
logic which reduce the time required to draw the minimap.