Fix a potential crash when changing to fish form with huggies attached.
This patch corrects a bug which can cause the game to crash when Naija
changes to fish form while huggies are attached, because the script
tries to use an already-freed bone pointer.  Another symptom is that the
huggies suddenly disappear when Naija changes form (due to being warped
to invalid coordinates), rather than flying off as usual.