Use HUGE_VALF insted of -1 as the default value for "smallest" searches.
This patch rewrites "find the smallest value" "find the closest object"
loops to use the HUGE_VALF constant, changing the following construct:

    float closest = -1;
    Object *closestObject = 0;
    for (Object *object = firstObject(); object; object = nextObject()) {
        float dist = distanceToObject(object);
        if (closest == -1 || dist < closest)
            closest = dist, closestObject = object;
    }

by initializing the distance to the currently closest object to infinity
instead of a negative value.  (The constant HUGE_VALF is normally
defined in <math.h> as a single-precision infinity, but a fallback is
supplied in BBGE/Base.h for environments which are missing the constant.)
This allows the "closest == -1" test to be eliminated; depending on the
loop structure and execution architecture, removing that single
comparison can reduce execution time by several percent.

As a further optimization, loops which include a test for a maximum
value (such as "find the closest object within 100 units") have that
test merged into the distance test by initializing the closest-distance
variable to that limit instead of HUGE_VALF.

In addition, some loops which previously used integers to hold the
current and smallest distances have been changed to use floats, since
it is typically less expensive to directly compare two floating-point
values than to convert from floating-point to integer and then compare
two integers.