Move script constants from entityinclude.lua to ScriptInterface.cpp.
This patch moves the global constants defined in
scripts/entities/entityinclude.lua into Aquaria/ScriptInterface.cpp, and
removes the dofile() call for entityinclude.lua from all scripts that
had such a call, in order to avoid taking the time to load it for every
script instance.  For parallelism and convenience, the list of functions
exported to scripts is also moved to an array, as opposed to using a
long sequence of function calls.

This patch also removes a few duplicate or conflicting definitions of
constants, leaving only the effective value in the table.