Add a header file for Aquaria/Protect.cpp.
This changeset adds a header file "Protect.h" for the source file
Aquaria/Protect.cpp, so the declaration of the getRegistrationName()
function defined in that file can be properly shared with DSQ.cpp.
This is only for completeness (avoiding potential bugs if the function
signature is ever changed) and has no effect on program behavior.