Allow songs to be checked on the song menu using joystick/keyboard controls.
This patch enables the use of the menu directional controls (e.g. arrow
keys or joystick directional pad) to highlight song icons in the song
menu, so the user is not required to use the mouse to do so.  The patch
also changes the song menu behavior so that either the left or right
mouse buttons (or any keyboard/joystick controls set for the mouse
button equivalent actions), not just the left mouse button, will trigger
Naija's description of the song.