Use a bitmap instead of a list for tracking visited areas of the map.
This patch changes the WorldMapTile class to use a bitmap rather than a
list of IntPairs for tracking which areas of the map have been visited
(the MAPVIS feature), and cleans up visited data handling.  The use of
a bitmap reduces worst-case memory usage (all areas of all map tiles
visited) by a factor of 128.

This change also resolves a bug in the current code which causes the
visited lists to grow in size every time the game is saved, since the
lists are not cleared at that time.  As a result, uncompressed save data
sizes could grow by an order of magnitude or more over the minimum
necessary size, depending on how frequently the player saved their game.
(This would in turn cause the game to freeze for a short time while
displaying a save slot page with one or more of these ballooned files.)
Any save files suffering from this bug can be fixed by loading and then
saving the game after applying this patch; a typical compressed save
file (*.aqs) from the endgame should be around 40-50k.

In addition, this patch adds an alternate format for the visited-area
data stored in save files.  Rather than a textual list of coordinate
pairs, this alternate format writes the data as a base64-encoded string
containing the raw bitmap data stored for each world map tile.  This
has the advantage of only requiring a small, fixed amount of memory to
decode, and can be parsed much more quickly than reading and converting
a long stream of numbers from a string.

The alternate format is enabled by defining the preprocessor macro
AQUARIA_SAVE_MAPVIS_RAW, which is undefined by default.

IMPORTANT:  As noted in the DSQ.h comments, save files generated with
this feature enabled WILL BE INCOMPATIBLE with builds of Aquaria which
do not include this patch!  This feature is intended primarily for
portable systems and other platforms where storage space or memory are
at such a premium that the current text-based format (potentially up to
1.3MB, with three copies required in memory) is impractical.