Fix showBuffer() timing in ScreenTransition::capture().
Move the showBuffer() call after the glCopyTexSubImage2D() call in
ScreenTransition::capture().  While this method is overridden by
AquariaScreenTransition::capture() in Aquaria/DSQ.cpp (which already
uses the proper timing), showBuffer() will swap the front and back
buffers and potentially cause the transition texture to be grabbed from
the wrong buffer, so I've fixed it for completeness.