Allow option menu sliders to be adjusted with the joystick.
This patch adds logic to the volume sliders on the options menu so that
the sliders can be adjusted using left and right on the joystick (or
directional pad, if supported).  As a side effect, this prevents using
the joystick to move directly from the sliders to the checkboxes on the
right side of the menu, though it's still possible to get there by going
down and around through the Apply and Cancel options.