Avoid using floating-point arithmetic to save and restore color/alpha values.
The Entity class calls RenderObject::multiplyColor() and multiplyAlpha()
to first multiply the color and alpha of all sprite bone objects by
those of the entity, and then restore the color and alpha by dividing by
the same values.  This is unreliable because:

(1) Multiplication and division are not guaranteed to be exact inverses
    of each other in the floating-point domain.  In particular, the x87
    floating-point instructions on the x86 architecture do not maintain
    precision as specified by IEEE 754, so even seemingly innocuous code
    like "a = b; assert(a == b);" may fail (see, for example,
    http://gcc.gnu.org/bugzilla/show_bug.cgi?id=323).

(2) If the entity's color or alpha are ever set to zero, the color or
    alpha of all component RenderObjects will be permanently lost.

For Aquaria's purposes, it is never necessary to stack two or more such
color/alpha modifiers, so this patch changes the code to simply save and
restore the original color and alpha values (at a cost of 16 extra bytes
of memory per RenderObject; since the number of RenderObjects is
typically no more than a few thousand, this should not be a significant
problem).