Fix a duplicate call of addRenderObject() on the same object during the intro sequence.
This patch fixes a typo in the intro sequence which caused the same
RenderObject to be added twice to a RenderObjectLayer, potentially
causing a crash when the object is subsequently deleted (and leaking
memory on the un-added object).

This naturally makes a difference in how the intro sequence looks, in
this case the large cloud that covers the screen at the very beginning
of the sequence, so I've commented out the second object entirely in
this patch to preserve (or at least stay as close as possible to)
current behavior.