Make sure the minimap stays hidden inside RADARHIDE nodes.
This patch fixes the minimap rendering logic so that the minimap is
always hidden inside a PATH_RADARHIDE node, even if the center of the
node containing Naija is farther away than that of another RADARHIDE
node.  This can be seen in mithalas01 room 8 (the house closest to the
palace entrance), in which the minimap would reappear if Naija moved to
the very bottom of the room.

This was reported on the Bit Blot forum, at:
http://www.bit-blot.com/forum/index.php?topic=1709.0