Fix missing return values from Lua interface functions.
This patch fixes a few of the Lua interface functions in
ScriptInterface.cpp which claimed to return values but did not actually
push values onto the stack; in doing so, it also corrects some invalid
Lua API calls (which were presumably typos).  While none of the return
values from these functions seem to be used in Lua code, this keeps the
function interface consistent.