Fixed crash bug with the sun form (found by Sindhi).
Explanation cut-and-pasted from mailing list:

"Reproduced it by starting up at the final save point, and then staying
idle in sun form for ~20 secs. Crash.

The problem is that these 2 calls ...

lightFormGlow->moveToFront();
lightFormGlowCone->moveToFront();

...do a move front all the time, alternating, leaving a lot of NULL
vector entries behind, until "firstFreeIdx = curIdx" reaches the end of
the vector (preallocated with 100), and overflows the index in the while
loop pointed out earlier.
As this happens only every 0.5s, it takes a while to actually reach the end.

The following snippet fixes it, but i suppose it is too hackish.
Is the vector ever compacted? Seems to me that RenderObjectLayer::sort()
is never actually called, and it does not set firstFreeIdx accordingly
anyway - it should, imho. If it was called."